UCT/IB: Make num_paths unsigned instead of uint8_t
Align the type of num_paths with path_bits_count and roce_path_factor

+++ src/uct/ib/base/ib_iface.h
@@ -189,7 +189,7 @@ struct uct_ib_iface {
 
     uint8_t                   *path_bits;
     unsigned                  path_bits_count;
-    uint8_t                   num_paths;
+    unsigned                  num_paths;
     uint16_t                  pkey_index;
     uint16_t                  pkey_value;
     uint8_t                   addr_size;
